+++ errors/errors.go
@@ -1,0 +1,2 @@
+// Package errors extends the errors package in the stdlib.
+package errors