+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- **errors:** `Recover` function to create an error from a recovered panic
+- **errors:** `Handler` interface to handle errors
+- **errors:** `LogHandler` implementation (tested with [sirupsen/logrus](https://github.com/sirupsen/logrus))
+- **errors:** `TestHandler` implementation for testing purposes
+- **errors:** `NullHandler` implementation as a fallback
+- **util:** `ShutdownManager` to register and execute shutdown handlers when an application exits
+
+
 ## 0.0.5 - 2017-05-10
 
 ### Added